+++ src/Config.hh
@@ -157,6 +157,12 @@ public:
 	CurrHeadSelector getCurrHeadSelector(void) const {
 		return _screen_curr_head_selector;
 	}
+	bool isDefaultFontX11(void) const {
+		return _screen_default_font_x11;
+	}
+	const std::string &getFontCharsetOverride(void) const {
+		return _screen_font_charset_override;
+	}
 
 	bool placeTransOnParent(void) const { return _place_trans_parent; }
 
@@ -306,6 +312,11 @@ private:
 	bool _screen_honour_aspectratio;
 	/** Setting for how current head is determined. */
 	CurrHeadSelector _screen_curr_head_selector;
+	/** If true, default font is X11 and not Xmb when no type
+	 * is specified in theme. */
+	bool _screen_default_font_x11;
+	/** Charset override for X11/Xmb fonts. */
+	std::string _screen_font_charset_override;
 	bool _screen_placement_row, _screen_placement_ltr, _screen_placement_ttb;
 	int _screen_placement_offset_x, _screen_placement_offset_y;
 	bool _place_trans_parent;