+++ src/FontHandler.cc
@@ -17,7 +17,9 @@
 
 static Util::StringTo<PFont::Type> map_type[] =
 	{{"X11", PFont::FONT_TYPE_X11},
+#ifdef PEKWM_HAVE_XFT
 	 {"XFT", PFont::FONT_TYPE_XFT},
+#endif // PEKWM_HAVE_XFT
 	 {"XMB", PFont::FONT_TYPE_XMB},
 	 {nullptr, PFont::FONT_TYPE_NO}};
 
@@ -27,12 +29,13 @@ static Util::StringTo<FontJustify> map_justify[] =
 	 {"RIGHT", FONT_JUSTIFY_RIGHT},
 	 {nullptr, FONT_JUSTIFY_NO}};
 
-//! @brief FontHandler constructor
-FontHandler::FontHandler(void)
+FontHandler::FontHandler(bool default_font_x11,
+			 const std::string &charset_override)
+	: _default_font_x11(default_font_x11),
+	  _charset_override(charset_override)
 {
 }
 
-//! @brief FontHandler destructor
 FontHandler::~FontHandler(void)
 {
 	std::vector<HandlerEntry<PFont*> >::iterator fit = _fonts.begin();
@@ -40,18 +43,20 @@ FontHandler::~FontHandler(void)
 		delete fit->getData();
 	}
 
-	std::vector<HandlerEntry<PFont::Color*> >::iterator cit = _colors.begin();
+	std::vector<HandlerEntry<PFont::Color*> >::iterator cit =
+		_colors.begin();
 	for (; cit != _colors.end(); ++cit) {
 		delete cit->getData();
 	}
 }
 
-//! @brief Gets or allocs a font
-//!
-//! Syntax of font specification goes as follows:
-//!   "Font Name#Justify#Offset#Type" ex "Vera#Center#1 1#XFT"
-//! where only the first field is obligatory and type needs to be the last
-//!
+/**
+ * @brief Gets or allocs a font
+ *
+ * Syntax of font specification goes as follows:
+ *   "Font Name#Justify#Offset#Type" ex "Vera#Center#1 1#XFT"
+ * where only the first field is obligatory and type needs to be the last
+ */
 PFont*
 FontHandler::getFont(const std::string &font)
 {
@@ -65,66 +70,19 @@ FontHandler::getFont(const std::string &font)
 	}
 
 	// create new
-	PFont *pfont = 0;
-
+	PFont::Type type;
 	std::vector<std::string> tok;
-	std::vector<std::string>::iterator tok_it;
-	if ((Util::splitString(font, tok, "#", 0, true)) > 1) {
-		// Try getting the font type from the first paramter, if that
-		// doesn't work fall back to the last. This is to backwards
-		// compatible.
-		tok_it = tok.begin();
-		uint type = Util::StringToGet(map_type, *tok_it);
-		if (type == PFont::FONT_TYPE_NO) {
-			tok_it = tok.end() - 1;
-			type = Util::StringToGet(map_type, *tok_it);
-		}
+	PFont *pfont = newFont(font, tok, type);
 
-		switch (type) {
-		case PFont::FONT_TYPE_XMB:
-			pfont = new PFontXmb;
-			tok.erase(tok_it);
-			break;
-#ifdef PEKWM_HAVE_XFT
-		case PFont::FONT_TYPE_XFT:
-			pfont = new PFontXft;
-			tok.erase(tok_it);
-			break;
-#endif // PEKWM_HAVE_XFT
-		case PFont::FONT_TYPE_X11:
-			pfont = new PFontX11;
-			tok.erase(tok_it);
-			break;
-		default:
-			pfont = new PFontXmb;
-			break;
-		};
-		pfont->load(tok.front());
-
-		// Remove used fields, type and
-		tok.erase(tok.begin());
-
-		// fields left for justify and offset
-		std::vector<std::string>::iterator s_it(tok.begin());
-		for (; s_it != tok.end(); ++s_it) {
-			if (isdigit((*s_it)[0])) { // number
-				std::vector<std::string> tok_2;
-				if (Util::splitString(*s_it, tok_2, " \t", 2) == 2) {
-					pfont->setOffset(strtol(tok_2[0].c_str(), 0, 10),
-							 strtol(tok_2[1].c_str(), 0, 10));
-				}
-			} else { // justify
-				uint justify = Util::StringToGet(map_justify, *s_it);
-				if (justify == FONT_JUSTIFY_NO) {
-					justify = FONT_JUSTIFY_LEFT;
-				}
-				pfont->setJustify(justify);
-			}
-		}
-	} else {
-		pfont = new PFontXmb;
-		pfont->load(font);
+	std::string font_str(tok.front());
+	tok.erase(tok.begin());
+	if (type == PFont::FONT_TYPE_X11 || type == PFont::FONT_TYPE_XMB) {
+		replaceFontCharset(font_str);
 	}
+	pfont->load(font_str);
+
+	// fields left for justify and offset
+	parseFontOptions(pfont, tok);
 
 	// create new entry
 	HandlerEntry<PFont*> entry(font);
@@ -134,6 +92,105 @@ FontHandler::getFont(const std::string &font)
 	_fonts.push_back(entry);
 
 	return pfont;
+}
+
+PFont*
+FontHandler::newFont(const std::string &font,
+		     std::vector<std::string> &tok, PFont::Type &type)
+{
+	std::vector<std::string>::iterator tok_it;
+	if (Util::splitString(font, tok, "#", 0, true) > 1) {
+		// Try getting the font type from the first paramter,
+		// if that doesn't work fall back to the last. This is to
+		// backwards compatible.
+		tok_it = tok.begin();
+		type = Util::StringToGet(map_type, *tok_it);
+		if (type == PFont::FONT_TYPE_NO) {
+			tok_it = tok.end() - 1;
+			type = Util::StringToGet(map_type, *tok_it);
+		}
+		if (type != PFont::FONT_TYPE_NO) {
+			tok.erase(tok_it);
+		}
+	} else {
+		type = PFont::FONT_TYPE_NO;
+	}
+
+	switch (type) {
+#ifdef PEKWM_HAVE_XFT
+	case PFont::FONT_TYPE_XFT:
+		return new PFontXft;
+#endif // PEKWM_HAVE_XFT
+	case PFont::FONT_TYPE_X11:
+		return new PFontX11;
+	case PFont::FONT_TYPE_XMB:
+		return new PFontXmb;
+	default:
+		if (_default_font_x11) {
+			return new PFontX11;
+		}
+		return new PFontXmb;
+	};
+}
+
+void
+FontHandler::parseFontOptions(PFont *pfont,
+			      std::vector<std::string> &tok)
+{
+	std::vector<std::string>::iterator s_it(tok.begin());
+	for (; s_it != tok.end(); ++s_it) {
+		if (! parseFontOffset(pfont, *s_it)) {
+			parseFontJustify(pfont, *s_it);
+		}
+	}
+}
+
+bool
+FontHandler::parseFontOffset(PFont *pfont, const std::string &str)
+{
+	if (str.size() == 0 || ! isdigit(str[0])) {
+		return false;
+	}
+
+	std::vector<std::string> tok;
+	if (Util::splitString(str, tok, " \t", 2) == 2) {
+		try {
+			pfont->setOffset(std::stoi(tok[0]),
+					 std::stoi(tok[1]));
+		} catch (std::invalid_argument&) { }
+	}
+	return true;
+}
+
+bool
+FontHandler::parseFontJustify(PFont *pfont, const std::string &str)
+{
+	uint justify = Util::StringToGet(map_justify, str);
+	if (justify == FONT_JUSTIFY_NO) {
+		justify = FONT_JUSTIFY_LEFT;
+	}
+	pfont->setJustify(justify);
+	return true;
+}
+
+/**
+ * Replace last two -*-* parts of a font string and replace with
+ * _charset_override if _charset_override is non empty.
+ */
+bool
+FontHandler::replaceFontCharset(std::string &font)
+{
+	size_t begin;
+	if (_charset_override.size() == 0
+	    || (begin = font.rfind('-')) == std::string::npos || (begin == 0)
+	    || (begin = font.rfind('-', begin - 1)) == std::string::npos) {
+		return false;
+	}
+
+	// keep the -, default charset should be specified without leading -
+	begin += 1;
+	font.replace(begin, font.size() - begin, _charset_override);
+	return true;
 }
 
 //! @brief Returns a font