+++ src/FontHandler.hh
@@ -20,8 +20,16 @@
 //! @brief FontHandler, a caching and font type transparent font handler.
 class FontHandler {
 public:
-	FontHandler(void);
+	FontHandler(bool default_font_x11,
+		    const std::string &charset_override);
 	~FontHandler(void);
+
+	void setDefaultFontX11(bool default_font_x11) {
+		_default_font_x11 = default_font_x11;
+	}
+	void setCharsetOverride(const std::string &charset_override) {
+		_charset_override = charset_override;
+	}
 
 	PFont *getFont(const std::string &font);
 	void returnFont(PFont *font);
@@ -29,10 +37,24 @@ public:
 	PFont::Color *getColor(const std::string &color);
 	void returnColor(PFont::Color *color);
 
-private:
-	void loadColor(const std::string &color, PFont::Color *font_color, bool fg);
+protected:
+	bool parseFontOffset(PFont *pfont, const std::string &str);
+	bool parseFontJustify(PFont *pfont, const std::string &str);
+	bool replaceFontCharset(std::string &font);
 
 private:
+	PFont *newFont(const std::string &font,
+		       std::vector<std::string> &tok, PFont::Type &type);
+	void parseFontOptions(PFont *pfont,
+			      std::vector<std::string> &tok);
+	void loadColor(const std::string &color, PFont::Color *font_color,
+		       bool fg);
+
+private:
+	/** If true, unspecified font type is X11 and not XMB. */
+	bool _default_font_x11;
+	/** If non empty, override charset in X11/XMB font strings. */
+	std::string _charset_override;
 	std::vector<HandlerEntry<PFont*> > _fonts;
 	std::vector<HandlerEntry<PFont::Color*> > _colors;
 };