+++ src/Globals.cc
@@ -77,7 +77,9 @@ namespace pekwm
 		_key_grabber->load(_config->getKeyFile());
 		_key_grabber->grabKeys(X11::getRoot());
 
-		_font_handler = new FontHandler();
+		_font_handler =
+			new FontHandler(_config->isDefaultFontX11(),
+					_config->getFontCharsetOverride());
 		_image_handler = new ImageHandler();
 		_texture_handler = new TextureHandler();
 		_theme = new Theme(_font_handler, _image_handler, _texture_handler,