+++ NEWS.md
@@ -12,6 +12,9 @@ New
   _pointer_, _workspacelastofocused_, _top_ and _root_.
 * **--standalone** option to _pekwm_wm_, convenience to improve the debugging
   experience.
+* **FontDefaultX11** and **FontCharsetOverride** options added to the _Screen_
+  section of the main configuration file. Controls the default font type and
+  override of charset in font strings.
 
 Updated
 -------